# shared/contracts/api.yaml
openapi: 3.1.0
info:
  title: AI-Rookie API
  version: 0.2.0
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  service:
                    type: string
                  storage:
                    $ref: "#/components/schemas/StorageStatus"
                  source:
                    $ref: "#/components/schemas/LawSourceStatus"
                required:
                  - status
                  - service
                  - storage
                  - source
  /source-status:
    get:
      summary: Report status for a request-selected law source provider
      parameters:
        - in: query
          name: provider
          schema:
            type: string
            enum:
              - local-fixture
              - law-go-public
              - korea-law-mcp
          description: Optional source provider to inspect. Defaults to the runtime default provider.
      responses:
        "200":
          description: Selected source-provider status
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestedProvider:
                    type: string
                  source:
                    $ref: "#/components/schemas/LawSourceStatus"
                required:
                  - requestedProvider
                  - source
  /source-search:
    get:
      summary: Search ordinance candidates for a request-selected law source provider
      parameters:
        - in: query
          name: provider
          schema:
            type: string
            enum:
              - local-fixture
              - law-go-public
              - korea-law-mcp
          description: Optional source provider to search against. Defaults to the runtime default provider.
        - in: query
          name: query
          required: true
          schema:
            type: string
          description: Ordinance title or keyword query.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 10
      responses:
        "200":
          description: Search candidates for the selected source provider
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestedProvider:
                    type: string
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/LawSourceSearchResult"
                  recommendation:
                    oneOf:
                      - $ref: "#/components/schemas/LawSourceSearchRecommendation"
                      - type: "null"
                  meta:
                    type: object
                    additionalProperties: true
                required:
                  - requestedProvider
                  - query
                  - results
                  - recommendation
                  - meta
  /case-catalog:
    get:
      summary: List bundled local-fixture case packs for the demo dashboard
      responses:
        "200":
          description: Bundled local case catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                    enum:
                      - local-fixture
                  defaultCaseId:
                    type: string
                  cases:
                    type: array
                    items:
                      $ref: "#/components/schemas/LocalFixtureCase"
                required:
                  - provider
                  - defaultCaseId
                  - cases
  /analyze:
    post:
      summary: Analyze regulation changes and suggest follow-up actions
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResponse"
        "400":
          description: Invalid request payload or JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Pipeline input validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /history:
    get:
      summary: List recent analysis runs saved to the configured storage backend
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: Recent analysis runs and storage status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
components:
  schemas:
    StorageStatus:
      type: object
      properties:
        enabled:
          type: boolean
        provider:
          type: string
          enum:
            - local
            - supabase
            - municipal
            - unsupported
        projectRef:
          type:
            - string
            - "null"
        missingEnv:
          type: array
          items:
            type: string
      required:
        - enabled
        - provider
        - missingEnv
    Clause:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        text:
          type: string
      required:
        - id
        - title
        - text
    RegulationDocument:
      type: object
      properties:
        clauses:
          type: array
          items:
            $ref: "#/components/schemas/Clause"
      required:
        - clauses
    InternalDocument:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        text:
          type: string
      required:
        - id
        - type
        - title
        - text
    AnalyzeRequest:
      type: object
      additionalProperties: false
      properties:
        source:
          $ref: "#/components/schemas/AnalyzeSourceRequest"
        before:
          $ref: "#/components/schemas/RegulationDocument"
        after:
          $ref: "#/components/schemas/RegulationDocument"
        internalDocs:
          type: array
          items:
            $ref: "#/components/schemas/InternalDocument"
    AnalyzeSourceRequest:
      type: object
      properties:
        provider:
          type: string
          enum:
            - local-fixture
            - law-go-public
            - korea-law-mcp
        beforeId:
          type: string
        afterId:
          type: string
        caseId:
          type: string
      description: >
        For `local-fixture`, `caseId` selects one of the bundled municipality case packs.
        For `law-go-public`, `beforeId` and `afterId` are ordinance sequence IDs from law.go.kr.
        For `korea-law-mcp`, `beforeId` and `afterId` are passed to the configured MCP detail tool.
    LawSourceStatus:
      type: object
      properties:
        enabled:
          type: boolean
        provider:
          type: string
          enum:
            - local-fixture
            - law-go-public
            - korea-law-mcp
            - unsupported
        mode:
          type: string
        missingEnv:
          type: array
          items:
            type: string
        transport:
          type: string
        ocMode:
          type: string
        baseUrl:
          type: string
        detailToolNames:
          type: array
          items:
            type: string
        idArgumentName:
          type: string
        searchToolNames:
          type: array
          items:
            type: string
        searchQueryArgumentName:
          type: string
        caseCount:
          type: integer
        defaultCaseId:
          type: string
      required:
        - enabled
        - provider
        - mode
        - missingEnv
    LawSourceSearchResult:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        jurisdiction:
          type: string
        effectiveDate:
          type: string
        promulgationDate:
          type: string
        referenceUrl:
          type: string
        summary:
          type: string
      required:
        - id
        - title
    LawSourceSearchRecommendation:
      type: object
      properties:
        before:
          $ref: "#/components/schemas/LawSourceSearchResult"
        after:
          $ref: "#/components/schemas/LawSourceSearchResult"
        confidence:
          type: string
          enum:
            - high
            - medium
            - low
        matchCount:
          type: integer
        reason:
          type: string
        strategy:
          type: string
      required:
        - before
        - after
        - confidence
        - matchCount
        - reason
        - strategy
    LocalFixtureCase:
      type: object
      properties:
        caseId:
          type: string
        title:
          type: string
        municipality:
          type: string
        domain:
          type: string
        effectiveDate:
          type: string
        ordinanceNo:
          type: string
        officialUrl:
          type: string
        normalizedDemoExcerpts:
          type: boolean
        defaultSample:
          type: boolean
      required:
        - caseId
        - title
        - municipality
        - domain
        - effectiveDate
        - ordinanceNo
        - officialUrl
        - normalizedDemoExcerpts
        - defaultSample
    AnalyzeInputSource:
      type: object
      properties:
        provider:
          type: string
        mode:
          type: string
        caseId:
          type: string
        caseTitle:
          type: string
        municipality:
          type: string
        officialUrl:
          type: string
    AnalyzeResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            schemaVersion:
              type: string
            generatedAt:
              type: string
              format: date-time
            totalChanges:
              type: integer
            highRiskChangeCount:
              type: integer
            changeTypeBreakdown:
              type: object
              additionalProperties:
                type: integer
            storage:
              $ref: "#/components/schemas/StorageStatus"
            inputSource:
              $ref: "#/components/schemas/AnalyzeInputSource"
        analysis:
          type: object
          properties:
            changes:
              type: array
              items:
                type: object
            impactedDocuments:
              type: array
              items:
                type: object
            risks:
              type: array
              items:
                type: object
            traces:
              type: array
              items:
                type: object
        drafts:
          type: object
          additionalProperties:
            type: string
        changes:
          type: array
          items:
            type: object
        mapped:
          type: array
          items:
            type: object
        risks:
          type: array
          items:
            type: object
        traces:
          type: array
          items:
            type: object
      required:
        - drafts
    HistoryRun:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        source:
          type: string
        totalChanges:
          type: integer
        highRiskChangeCount:
          type: integer
        changeTypeBreakdown:
          type: object
          additionalProperties:
            type: integer
        result:
          $ref: "#/components/schemas/AnalyzeResponse"
      required:
        - id
        - createdAt
        - source
        - totalChanges
        - highRiskChangeCount
        - changeTypeBreakdown
        - result
    HistoryResponse:
      type: object
      properties:
        storage:
          $ref: "#/components/schemas/StorageStatus"
        runs:
          type: array
          items:
            $ref: "#/components/schemas/HistoryRun"
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
      required:
        - storage
        - runs
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
                required:
                  - path
                  - message
          required:
            - code
            - message
            - details
      required:
        - error